refactor: move utilities to cargo
